Build: Adjusting to file moves
