+++ configuration.conf
@@ -38,7 +38,6 @@ spanningTree        | Y | Y | Y
 mincut              | Y | Y | Y
 version             | Y | Y | Y
 transitiveClosure   | Y | Y | Y
-breadthFirstSearch  | Y | Y | N
 traversal           | Y | Y | Y
 coloring            | Y | Y | Y
 planar              | Y | Y | Y
