+++ include/drivers/traversal/binaryBreadthFirstSearch_driver.h
@@ -27,8 +27,8 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
  ********************************************************************PGR-GNU*/
 
-#ifndef INCLUDE_DRIVERS_BREADTHFIRSTSEARCH_BINARYBREADTHFIRSTSEARCH_DRIVER_H_
-#define INCLUDE_DRIVERS_BREADTHFIRSTSEARCH_BINARYBREADTHFIRSTSEARCH_DRIVER_H_
+#ifndef INCLUDE_DRIVERS_TRAVERSAL_BINARYBREADTHFIRSTSEARCH_DRIVER_H_
+#define INCLUDE_DRIVERS_TRAVERSAL_BINARYBREADTHFIRSTSEARCH_DRIVER_H_
 #pragma once
 
 
@@ -73,4 +73,4 @@ void pgr_do_binaryBreadthFirstSearch(
 }
 #endif
 
-#endif  // INCLUDE_DRIVERS_BREADTHFIRSTSEARCH_BINARYBREADTHFIRSTSEARCH_DRIVER_H_
+#endif  // INCLUDE_DRIVERS_TRAVERSAL_BINARYBREADTHFIRSTSEARCH_DRIVER_H_
