+++ sql/breadthFirstSearch/CMakeLists.txt
@@ -3,9 +3,7 @@
 # License: GPL-2 See https://github.com/pgRouting/pgrouting/blob/main/LICENSE
 
 set(LOCAL_FILES
-	_breadthFirstSearch.sql
 	_binaryBreadthFirstSearch.sql
-    breadthFirstSearch.sql
     binaryBreadthFirstSearch.sql
     )
