+++ src/traversal/CMakeLists.txt
@@ -3,5 +3,9 @@
 # License: GPL-2 See https://github.com/pgRouting/pgrouting/blob/main/LICENSE
 
 ADD_LIBRARY(traversal OBJECT
+    breadthFirstSearch.c
     depthFirstSearch.c
+    binaryBreadthFirstSearch.c
+
+    binaryBreadthFirstSearch_driver.cpp
 )
