+++ src/traversal/binaryBreadthFirstSearch.c
@@ -35,7 +35,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include "c_common/debug_macro.h"
 #include "c_common/e_report.h"
 #include "c_common/time_msg.h"
-#include "drivers/breadthFirstSearch/binaryBreadthFirstSearch_driver.h"
+#include "drivers/traversal/binaryBreadthFirstSearch_driver.h"
 
 PGDLLEXPORT Datum _pgr_binarybreadthfirstsearch(PG_FUNCTION_ARGS);
 PG_FUNCTION_INFO_V1(_pgr_binarybreadthfirstsearch);
